DatabaseManager constructs RocksResources from RocksDbConfig
Move `RocksDbConfig` (with its `ByteSize` fields and defaults) from
`server::parameters::config` down to `common/options` so both `server`
and `database` can depend on it without crossing layers.

Change `DatabaseManager::new` to take `&RocksDbConfig` instead of a
pre-built `Arc<RocksResources>` and build the shared cache and write-
buffer manager itself. The server entry point now just hands its
storage config to the database manager:

    DatabaseManager::new(&config.storage.data_directory,
                         &config.storage.rocksdb)

The "Storage configured: rocksdb cache=..., write-buffers-limit=..."
debug log moves with it, so the message is emitted at the natural
construction site instead of two layers up.

Test/bench call sites that previously passed an explicit 64 MB / 64 MB
`RocksResources` now pass `&RocksDbConfig::default()` (1 GiB cache,
512 MiB WBM). The old per-test small budgets were never actually
allocated up front — RocksDB grows the cache lazily — so the larger
defaults don't affect peak memory of unit tests.
