+++ common/options/options.rs
@@ -6,7 +6,32 @@
 
 pub mod byte_size;
 
+use serde::Deserialize;
+
 pub use byte_size::{ByteSize, ParseByteSizeError};
+
+#[derive(Clone, Debug, Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub struct RocksDbConfig {
+    #[serde(default = "default_rocksdb_cache_size")]
+    pub cache_size: ByteSize,
+    #[serde(default = "default_rocksdb_write_buffers_limit")]
+    pub write_buffers_limit: ByteSize,
+}
+
+impl Default for RocksDbConfig {
+    fn default() -> Self {
+        Self { cache_size: default_rocksdb_cache_size(), write_buffers_limit: default_rocksdb_write_buffers_limit() }
+    }
+}
+
+const fn default_rocksdb_cache_size() -> ByteSize {
+    ByteSize::gb(1)
+}
+
+const fn default_rocksdb_write_buffers_limit() -> ByteSize {
+    ByteSize::mb(512)
+}
 
 use resource::constants::server::{
     DEFAULT_ANSWER_COUNT_LIMIT_GRPC, DEFAULT_ANSWER_COUNT_LIMIT_HTTP, DEFAULT_INCLUDE_INSTANCE_TYPES,
