+++ database/database_manager.rs
@@ -13,6 +13,7 @@ use std::{
 
 use cache::CACHE_DB_NAME_PREFIX;
 use diagnostics::diagnostics_manager::DiagnosticsManager;
+use options::RocksDbConfig;
 use resource::{constants::database::INTERNAL_DATABASE_PREFIX, internal_database_prefix};
 use storage::durability_client::WALClient;
 use storage::keyspace::storage_resources::RocksResources;
@@ -39,10 +40,19 @@ impl DatabaseManager {
     pub fn new(
         data_directory: impl AsRef<Path>,
         diagnostics_manager: Arc<DiagnosticsManager>,
-        rocks_resources: Arc<RocksResources>,
+        rocksdb_config: &RocksDbConfig,
     ) -> Result<Arc<Self>, DatabaseOpenError> {
         let data_directory = data_directory.as_ref().to_owned();
         let import_directory = data_directory.join(Self::IMPORT_DIRECTORY_NAME);
+
+        let rocks_resources = Arc::new(RocksResources::new(
+            rocksdb_config.cache_size.as_usize(),
+            rocksdb_config.write_buffers_limit.as_usize(),
+        ));
+        debug!(
+            "Storage configured: rocksdb cache={}, write-buffers-limit={}",
+            rocksdb_config.cache_size, rocksdb_config.write_buffers_limit,
+        );
 
         let databases = RwLock::new(Self::initialise_databases(
             &data_directory,
